updated landing page content
