+++ app/components/ui/icons/name.d.ts
@@ -14,6 +14,7 @@ export type IconName =
 	| 'exit'
 	| 'file-text'
 	| 'github-logo'
+	| 'instagram-logo'
 	| 'laptop'
 	| 'link-2'
 	| 'lock-closed'
